chore: remove redundant GitHub workflow files and custom build steps
Removed multiple GitHub workflow files including Build.yml, BuildPrivateNugetFeed.yml, Cleanup.yml, and Test_BuildWithCustomArtifacts.yml. Also removed CustomStep.cs containing custom build steps. These deletions are aimed at cleaning up outdated or unused workflow configurations.
