+++ .devops/workflows/CleanupBuilds.yml
@@ -1,0 +1,22 @@
+name: CleanupBuilds
+
+variables:
+  - group: Atom
+trigger:
+  branches:
+  tags:
+    include:
+      - 'v[0-9]+.[0-9]+.[0-9]+'
+
+jobs:
+  
+  - job: CleanupPrereleaseArtifacts
+    pool:
+      vmImage: ubuntu-latest
+    steps:
+      
+      - checkout: self
+        fetchDepth: 0
+      
+      - script: dotnet run --project _atom/_atom.csproj CleanupPrereleaseArtifacts --skip --headless
+        name: CleanupPrereleaseArtifacts
