+++ .github/workflows/Build.yml
@@ -230,26 +230,6 @@ jobs:
           path: "${{ github.workspace }}/.github/publish/DecSm.Atom.Tests"
       
   
-  TestPrivateNugetRestore:
-    runs-on: ubuntu-latest
-    steps:
-      
-      - name: Checkout
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-      
-      - name: Setup NuGet
-        run: |
-          dotnet run --project DecSm.Atom.Tools.Nuget/DecSm.Atom.Tools.Nuget.csproj -- nuget-add --feed "DecSm;https://nuget.pkg.github.com/DecSmith42/index.json"
-        shell: bash
-        env:
-          NUGET_TOKEN_DECSM: ${{ secrets.PRIVATE_NUGET_API_KEY }}
-        
-      - name: TestPrivateNugetRestore
-        id: TestPrivateNugetRestore
-        run: dotnet run --project _atom/_atom.csproj TestPrivateNugetRestore --skip --headless
-  
   PushToNuget:
     needs: [ PackAtom, PackAtomTool, PackAzureKeyVaultModule, PackAzureStorageModule, PackDevopsWorkflowsModule, PackDotnetModule, PackGithubWorkflowsModule, PackGitVersionModule ]
     runs-on: ubuntu-latest
@@ -316,29 +296,3 @@ jobs:
           azure-vault-tenant-id: ${{ secrets.AZURE_VAULT_TENANT_ID }}
           azure-vault-app-id: ${{ secrets.AZURE_VAULT_APP_ID }}
           azure-vault-app-secret: ${{ secrets.AZURE_VAULT_APP_SECRET }}
-  
-  PushToPrivateNuget:
-    needs: [ PackPrivateTestLib ]
-    runs-on: ubuntu-latest
-    steps:
-      
-      - name: Checkout
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-      
-      - name: Download PrivateTestLib
-        uses: actions/download-artifact@v4
-        with:
-          name: PrivateTestLib
-          path: "${{ github.workspace }}/.github/artifacts/PrivateTestLib"
-      
-      - name: PushToPrivateNuget
-        id: PushToPrivateNuget
-        run: dotnet run --project _atom/_atom.csproj PushToPrivateNuget --skip --headless
-        env:
-          azure-vault-address: ${{ secrets.AZURE_VAULT_ADDRESS }}
-          azure-vault-tenant-id: ${{ secrets.AZURE_VAULT_TENANT_ID }}
-          azure-vault-app-id: ${{ secrets.AZURE_VAULT_APP_ID }}
-          azure-vault-app-secret: ${{ secrets.AZURE_VAULT_APP_SECRET }}
-          private-nuget-api-key: ${{ secrets.PRIVATE_NUGET_API_KEY }}
