+++ .github/workflows/CleanupBuilds.yml
@@ -1,4 +1,4 @@
-name: Cleanup
+name: CleanupBuilds
 
 on:
   workflow_dispatch:
