+++ DecSm.Atom.Module.Dotnet/TransformProjectVersionScope.cs
