+++ DecSm.Atom.sln
@@ -102,5 +102,7 @@ Global
 		{15CEEDA9-E8B8-43D9-A3D2-7F93A97F17B2} = {5FFA3AFC-036E-49C2-B0CD-93DD5EDCEBAD}
 		{DF50CF84-BF79-403D-92EE-11DDABB5AE2B} = {5FFA3AFC-036E-49C2-B0CD-93DD5EDCEBAD}
 		{E4FD68F9-1828-4F29-A6A1-5B027B4CCC00} = {E9BC7470-34D0-4EDE-A8B9-DB40C785EBFB}
+		{BCF74F5E-9D9E-4DC2-9EEC-AAE9E0F4CB9D} = {5AD016BB-9A1F-4D16-AD1F-083184D7689E}
+		{3E0FA800-2C75-4E94-9756-58E5FA42867E} = {5AD016BB-9A1F-4D16-AD1F-083184D7689E}
 	EndGlobalSection
 EndGlobal
