+++ DecSm.Atom/Nuget/AddNugetFeedsStep.cs
@@ -1,0 +1,11 @@
+﻿using DecSm.Atom.Workflows.Options;
+
+namespace DecSm.Atom.Nuget;
+
+public sealed record AddNugetFeedsStep : CustomStep
+{
+    public IReadOnlyList<NugetFeedOptions> FeedsToAdd { get; init; } = [];
+
+    public static string EnvVar(string feedName) =>
+        $"NUGET_TOKEN_{feedName.Replace(" ", "_").ToUpper()}";
+}
