+++ DecSm.Atom/Workflows/Definition/IWorkflowOption.cs
@@ -30,9 +30,9 @@ public interface IWorkflowOption
 public abstract record WorkflowOption<TData, TSelf> : IWorkflowOption
     where TSelf : WorkflowOption<TData, TSelf>, new()
 {
-    public virtual bool AllowMultiple => false;
-
     public virtual TData? Value { get; init; }
+
+    public virtual bool AllowMultiple => false;
 
     public static TSelf Create(TData value) =>
         new()
