+++ DecSm.Atom/Workflows/Options/CustomStep.cs
@@ -1,0 +1,8 @@
+﻿namespace DecSm.Atom.Workflows.Options;
+
+public abstract record CustomStep : IWorkflowOption
+{
+    public string? Name { get; init; }
+
+    public bool AllowMultiple => true;
+}
