+++ _atom/_usings.cs
@@ -14,7 +14,9 @@ global using DecSm.Atom.Module.GithubWorkflows;
 global using DecSm.Atom.Module.GithubWorkflows.Generation.Options;
 global using DecSm.Atom.Module.GithubWorkflows.Triggers;
 global using DecSm.Atom.Module.GitVersion;
+global using DecSm.Atom.Nuget;
 global using DecSm.Atom.Params;
 global using DecSm.Atom.Process;
 global using DecSm.Atom.Workflows.Definition;
 global using DecSm.Atom.Workflows.Definition.Command;
+global using DecSm.Atom.Workflows.Options;
